+++ src/components/Layout/Layout.tsx
@@ -7,11 +7,11 @@ export default function Layout() {
   return (
     <div className="relative flex">
       <Modal />
-      <div className="absolute top-0 left-0 z-10">
-        <Sidebar />
-      </div>
-      <div className="absolute top-0 z-5">
+      <div className="absolute top-0 z-0">
         <Navbar />
+      </div>
+      <div className="absolute top-0 left-0">
+        <Sidebar />
       </div>
       <div className="tablet:mt-[7rem] mt-[6rem] ml-[6.7rem] tablet:ml-[16rem] desktop:ml-[30rem]">
         <Outlet />
